suzerain: silence -Wsign-compare warning
git-svn-id: https://svn.ices.utexas.edu/repos/pecos/suzerain/trunk@41249 0bd85f0b-c151-43b9-a1e9-27231104551f

+++ apps/reacting/quantities.cpp
@@ -256,7 +256,7 @@ bool quantities::load(const esio_handle h)
         if (this->Ns==0)
             { this->Ns = bglobal; }
         else
-            { assert(this->Ns==bglobal); }
+            { assert(static_cast<int>(this->Ns)==bglobal); }
 
         quantities_loader f("quantities", h, "bar_");
         f("rho_s", this->rho_s(0, Ns));
